// src/interface/interfaces.ts
export interface Movie {
    name: string;
    description: string;
    thumbnail: string;
}

export interface item {
    name: string;
    description: string;
    thumbnail: {
        path: string;
        extension: string;
    };  
}

export interface props{
    height?: number;
    src?: string;
}

export interface PropsProvider{
    children: JSX.Element | JSX.Element[]
}

export interface AppState{
    name: string;
    url: string;
    time: string;
}